extract contractEventParser common code
